requirements.txt
Flask
mysql-connector-python
python-dotenv
flask-cors
flask-limiter
bcrypt
PyJWT
bleach
